refac: refactor KZG library and add docs
