+++ kzg/src/lib.rs
@@ -1,59 +1,17 @@
 use ark_ec::{pairing::Pairing, Group};
-use ark_ff::{field_hashers::HashToField, fields::Field, PrimeField};
+use ark_ff::{field_hashers::HashToField, PrimeField};
 
 use polynomials::univariate_polynomial::UnivariatePolynomial;
 
-// For a detailed explanation on KZG, visit: https://blog.subspace.network/kzg-polynomial-commitments-cd64af8ec868
-
-// KZG
-
-// if f(z) = b
-// f(z) - b = 0
-// Based on this, z is a root of this polynomial (f(z) - b)
-// Therefore (x - z) is a factor of the polynomial f(z) - b
-// Abstracting the value of z, we have
-// f(x) - b
-// where (x -z) is a factor
-// Therefore, dividing f(x) - b by (x - z) should give us a polynomial without a remainder
-// f(x) - b / (x - z) = q(x)
-// where:
-// f(x) = the polynomial we want to commit to
-// x = can be any point we want to evaluate the polynomial at (in kzg it is used for the commitment and gotten from the trusted setup)
-// z = the point we want to open the polynomial at
-// b = the result of opening the polynomial at z
-// q = quotient polynoial (should have no remainder)
+mod trusted_setup;
+use trusted_setup::TrustedSetUpCeremony;
 
 pub struct KZG<C: Pairing, H: HashToField<C::ScalarField>> {
     trusted_setup: TrustedSetUpCeremony<C, H>,
 }
 
-// pub struct PolyCommitment<C: PrimeField > {
-//     eval: C,
-//     proof: C,
-//     quotient_poly: UnivariatePolynomial<C>
-// }
-
-// impl <C: PrimeField > PolyCommitment<C> {
-
-//     pub fn new(eval: C, proof: C, quotient_poly: UnivariatePolynomial<C>) -> Self {
-//         Self { eval, proof, quotient_poly }
-//     }
-// }
-
-// TODO:
-// pub fn generate_zerifier_for_nth_roots_of_unity<F: PrimeField>(domain_size: usize) -> UnivariatePolynomial<F> {
-//     // todo!()
-// }
-// pub fn generate_zerifier_for_domain<F: PrimeField>(domain_size: usize) -> UnivariatePolynomial<F> {
-//     // todo!()
-// }
-
 impl<C: Pairing, H: HashToField<C::ScalarField>> KZG<C, H> {
-    pub fn instantiate(
-        hasher_domain: &[u8],
-        domain_size: usize,
-        initial_randomness: &[u8],
-    ) -> Self {
+    fn instantiate(hasher_domain: &[u8], domain_size: usize, initial_randomness: &[u8]) -> Self {
         let trusted_setup_ceremony = TrustedSetUpCeremony::instantiate(
             hasher_domain,
             domain_size as u64,
@@ -65,11 +23,11 @@ impl<C: Pairing, H: HashToField<C::ScalarField>> KZG<C, H> {
         }
     }
 
-    pub fn contribute(&mut self, randomness: &[u8]) {
+    fn contribute(&mut self, randomness: &[u8]) {
         self.trusted_setup.contribute(randomness);
     }
 
-    pub fn commit_to_poly(&self, poly: &UnivariatePolynomial<C::ScalarField>) -> C::G1 {
+    fn commit_to_poly(&self, poly: &UnivariatePolynomial<C::ScalarField>) -> C::G1 {
         assert!(
             self.trusted_setup.public_parameters.g1_powers_of_tau.len() >= poly.coefficients.len(),
             "Powers of tau not sufficient to evaluate poly"
@@ -90,7 +48,7 @@ impl<C: Pairing, H: HashToField<C::ScalarField>> KZG<C, H> {
 
     // Opens a polynomial (poly) at a point (eval_point) using the commitment to the polynomial
     // returns a tuple of the evaluation and proof
-    pub fn open(
+    fn open(
         &self,
         eval_point: C::ScalarField,
         poly: UnivariatePolynomial<C::ScalarField>,
@@ -118,7 +76,7 @@ impl<C: Pairing, H: HashToField<C::ScalarField>> KZG<C, H> {
         (res, proof)
     }
 
-    pub fn verify(
+    fn verify(
         &self,
         eval_point: C::ScalarField,
         eval: C::ScalarField,
@@ -131,98 +89,6 @@ impl<C: Pairing, H: HashToField<C::ScalarField>> KZG<C, H> {
             - C::G2::generator().mul_bigint(eval_point.into_bigint());
 
         C::pairing(num, C::G2::generator()) == C::pairing(proof, denum)
-    }
-}
-
-//////////////////////////////////////////
-/// Trusted Setup Ceremony
-//////////////////////////////////////////
-
-#[derive(Debug)]
-pub struct PublicParameters<C: Pairing> {
-    // powers of tau for G1
-    pub g1_powers_of_tau: Vec<C::G1>,
-    // tau in G2
-    pub g2_power_of_tau: Vec<C::G2>,
-}
-
-impl<C: Pairing> PublicParameters<C> {
-    pub fn new(g1_powers_of_tau: Vec<C::G1>, g2_power_of_tau: Vec<C::G2>) -> Self {
-        Self {
-            g1_powers_of_tau,
-            g2_power_of_tau,
-        }
-    }
-}
-
-pub struct TrustedSetUpCeremony<C: Pairing, H: HashToField<C::ScalarField>> {
-    pub hasher: H,
-    pub domain_size: u64,
-    pub public_parameters: PublicParameters<C>, // Not considered for now
-                                                // pub contributors_proof: Vec<F>
-}
-
-// TODO: Review
-impl<C: Pairing, H: HashToField<C::ScalarField>> TrustedSetUpCeremony<C, H> {
-    // TODO:
-    // Correct domain size to degree
-    pub fn instantiate(hasher_domain: &[u8], domain_size: u64, randomness: &[u8]) -> Self {
-        let hasher = HashToField::new(hasher_domain);
-
-        let mut instance = Self {
-            hasher,
-            domain_size,
-            public_parameters: PublicParameters::new(vec![], vec![]),
-        };
-
-        let tau = instance
-            .hasher
-            .hash_to_field(randomness, 1)
-            .get(0)
-            .unwrap()
-            .to_owned();
-
-        let mut g1_powers_of_tau: Vec<C::G1> = vec![];
-
-        let g = C::G1::generator();
-
-        g1_powers_of_tau.push(g);
-
-        for i in 0..domain_size {
-            g1_powers_of_tau.push(g1_powers_of_tau[i as usize].mul_bigint(tau.into_bigint()));
-        }
-
-        let g2_power_of_tau = vec![C::G2::generator().mul_bigint(tau.pow([1]).into_bigint())];
-
-        let public_parameters = PublicParameters::new(g1_powers_of_tau, g2_power_of_tau);
-
-        instance.public_parameters = public_parameters;
-
-        instance
-    }
-
-    pub fn contribute(&mut self, randomness: &[u8]) {
-        let tau: C::ScalarField = self
-            .hasher
-            .hash_to_field(randomness, 1)
-            .get(0)
-            .unwrap()
-            .to_owned();
-
-        self.public_parameters.g1_powers_of_tau = self
-            .public_parameters
-            .g1_powers_of_tau
-            .iter()
-            .enumerate()
-            .map(|(index, val)| val.mul_bigint(tau.pow([index as u64]).into_bigint()))
-            .collect();
-
-        self.public_parameters.g2_power_of_tau = self
-            .public_parameters
-            .g2_power_of_tau
-            .iter()
-            .map(|val| val.mul_bigint(tau.into_bigint()))
-            .collect();
     }
 }
 
@@ -253,8 +119,6 @@ pub mod test {
         dbg!(trusted_setup_instance.public_parameters);
     }
 
-    // TODO
-    // Write a proper test
     #[test]
     pub fn test_commit_to_poly() {
         let poly = UnivariatePolynomial::new(vec![Fr::from(1), Fr::from(3), Fr::from(2)]);
